refactor(api): unify public interfaces in commands and server modules
Eliminate cross-package imports in commands/ and server/ modules by
consolidating external dependencies through package-level __init__.py
public interfaces.

Changes:
- commands/__init__.py: re-export _validate_pid_file from server
- commands/status.py: import _validate_pid_file from commands package
- server/__init__.py: re-export FailedGate, FlowManager from domain,
  CircuitBreaker, HeartbeatServer from runtime
- server/registry.py: import from server package instead of direct
  domain/runtime imports

Verification:
- mypy: Success (45 source files)
- pytest: 298 passed (commands), 10 passed (server)
- ruff: All checks passed

Ref: #1718
